+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
+## 1.3.0
+- Added Persistent SFX Audio Nodes.
+- Fixed audio looping bug.
+- Added events: isMouseDown, isMouseUp, isMousePressed, isKeyDown, isKeyPressed, getMouseWheel
+getMousePos
+- Added isHit method in DisplayObject in JoyDO.js
+
 ## 1.2.0
 - Added AudioInstance in JoyDO.js and Audio Support for the runtime.
 ## 1.1.0